requirements.txt
ansicolor
argparse
bs4
cx_Oracle
hashlib
pyjks
requests